+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.2] - 2023-06-13
+
+### Fixed
+
+- Storing items inside turtle via peripheral owner system
+
 ## [0.6.1] - 2023-06-13
 
 ### Fixed